requirements.txt
flask>=3.0.0
mollie-api-python>=3.5.0
gunicorn>=21.2.0
